Revert "Removed inf from op_registry generation and in all examples."
This reverts commit 3cc02f2e9f9fb82c935b34f69138f1c9c4f99cec.

+++ tests/layer/conv_fp16output/conv_fp16output.cpp
@@ -12,7 +12,7 @@ int main()
     std::vector<int64_t> weightsData0 = mv::utils::generateSequence<int64_t> (16, 1, 0);
 
     auto weights0 = om.constantInt(weightsData0,{1,1,16,1}, mv::DType("UInt8"), mv::Order::getRowMajorID(4), {{0},{1.0},{},{}});
-    auto conv0 = om.conv(input0, weights0, {1, 1}, {0, 0, 0, 0}, 1, 1,  mv::DType("UInt8"),{{0},{1.0},{},{}} , "conv");
+    auto conv0 = om.conv(input0, weights0, {1, 1}, {0, 0, 0, 0}, 1, 1,  mv::DType("Float16"),{{0},{1.0},{},{}} , "conv");
     om.output(conv0);
 
     std::string compDescPath = mv::utils::projectRootPath() + "/config/compilation/release_kmb_MC-Prefetch1.json";
